+++ src/arena/arena_write.c
@@ -8,9 +8,9 @@
 #include "my.h"
 #include "op.h"
 
-void arena_write(arena_t *arena, size_t pos, uint8_t n, size_t value)
+void arena_write(arena_t *arena, int64_t pos, uint8_t n, int64_t value)
 {
-    for (size_t i = 0; i < n; i++)
+    for (int64_t i = 0; i < n; i++)
         arena->arena_copy[(pos + i) % MEM_SIZE] = (value >> (2 * (n - i - 1)))
             & 0x03;
 }
