+++ src/process/process_get_arg_value.c
@@ -8,14 +8,14 @@
 #include "my.h"
 #include "op.h"
 
-size_t process_get_arg_value(process_t *process, arena_t *arena, size_t arg_id)
+int64_t process_get_arg_value(process_t *process, arena_t *arena, int64_t arg_id)
 {
-    size_t offset = 2;
-    size_t arg_type = 0;
+    int64_t offset = 2;
+    int64_t arg_type = 0;
 
     if (arg_id >= MAX_ARGS_NUMBER)
         return 0;
-    for (size_t i = 0; i < arg_id; i++)
+    for (int64_t i = 0; i < arg_id; i++)
         offset += process_get_arg_size(process, arena,
             process_get_arg_type(process, arena, i));
     arg_type = process_get_arg_type(process, arena, arg_id);
